+++ src/components/Select/MultiSelect.stories.tsx
@@ -37,17 +37,17 @@ export default {
   title: "Forms/MultiSelect",
   tags: ["form-field", "select", "autodocs"],
   argTypes: {
-    label: { control: "string" },
+    label: { control: "text" },
     disabled: { control: "boolean" },
-    error: { control: "string" },
-    value: { control: "string" },
-    defaultValue: { control: "string" },
+    error: { control: "text" },
+    value: { control: "text" },
+    defaultValue: { control: "text" },
     open: { control: "inline-radio", options: [undefined, true, false] },
     defaultOpen: { control: "boolean" },
-    name: { control: "string" },
+    name: { control: "text" },
     required: { control: "boolean" },
     showSearch: { control: "boolean" },
-    isFormControl: { control: "boolean" },
+    form: { control: "text" },
     clickableNoData: { control: "boolean" },
     showCheck: { control: "boolean" },
     orientation: { control: "inline-radio", options: ["horizontal", "vertical"] },
@@ -60,5 +60,46 @@ export const Playground = {
     label: "Label",
     value: "content1",
     showSearch: true,
+  },
+  parameters: {
+    docs: {
+      source: {
+        transform: (_: string, story: { args: Props; [x: string]: any }) => {
+          const { clickableNoData, showSearch, value, ...props } = story.args;
+          return `<MultiSelect\n  value={${JSON.stringify((value ?? "").split(","))}}\n${
+            clickableNoData
+              ? "  onCreateOption={search => console.log('Clicked ', search)}\n"
+              : ""
+          }${Object.entries(props)
+            .map(
+              ([key, value]) =>
+                `  ${key}=${typeof value == "string" ? `"${value}"` : `{${value}}`}`
+            )
+            .join("\n")}
+>
+  <MultiSelect.Trigger />
+  <MultiSelect.Content ${showSearch ? "showSearch" : ""}>
+    <MultiSelect.Group heading="Group label">
+      <MultiSelect.Item value="content0">
+        <Icon name="user" />
+        Content0
+      </MultiSelect.Item>
+    </MultiSelect.Group>
+    <div>
+      <MultiSelect.Item value="content1">Content1 long text content</MultiSelect.Item>
+    </div>
+    <MultiSelect.Item
+      value="content2"
+      disabled
+    >
+      Content2
+    </MultiSelect.Item>
+    <MultiSelect.Item value="content3">Content3</MultiSelect.Item>
+    <MultiSelect.NoData />
+  </MultiSelect.Content>
+</MultiSelect>`;
+        },
+      },
+    },
   },
 };
